+++ installers/app-installer/source/ftrack_app_installer/__init__.py
@@ -125,6 +125,12 @@ class WindowsAppInstaller(AppInstaller):
     codesign_folder = os.path.join(os_root_folder, "codesign")
     INNOSETUP_PATH = "C:\\Program Files (x86)\\Inno Setup 6\\ISCC.exe"
 
+    @property
+    def executable_path(self):
+        return os.path.join(
+            self.dist_path, self.bundle_name, f'{self.bundle_name}.exe'
+        )
+
     def codesign(self, path):
         '''Codesign artifact *path* using jsign tool in Windows'''
         bat_file = os.path.join(self.codesign_folder, "codesign.bat")
@@ -137,6 +143,9 @@ class WindowsAppInstaller(AppInstaller):
 
         if not os.path.exists(self.INNOSETUP_PATH):
             raise Exception(f'Inno Setup not found at: {self.INNOSETUP_PATH}')
+
+        if codesign:
+            self.codesign(self.executable_path)
 
         # Load template and inject data
         with open(